+++ client/scripts/db/bsd.js
@@ -96,5 +96,7 @@ module.exports = {
         value: 16
       }
     ],
-  }
+  },
+  memoryPosition: [
+  ],
 };