+++ client/scripts/db/highspeed2.js
@@ -103,5 +103,7 @@ module.exports = {
         value: 16
       }
     ],
-  }
+  },
+  memoryPosition: [
+  ],
 };