+++ client/scripts/db/ngg.js
@@ -106,5 +106,11 @@ module.exports = {
         value: 16
       }
     ],
-  }
+  },
+  memoryPosition: [
+    //{ offset: 0x326, name: 'TEXT', description: 'random visible text', type: 'string' },
+
+    { offset: 0x1CFA, name: 'CREDITS_FULL', description: '0-10 credits', type: 'uint8' },
+    { offset: 0x1CFB, name: 'CREDITS_HALF', description: '0: no half credits', type: 'uint8' },
+  ],
 };