+++ client/scripts/main.js
@@ -95,7 +95,7 @@ function initialiseEmu(gameEntry) {
 
           const { averageRTTms, sentMessages, failedMessages } = wpcEmuWebWorkerApi.getStatistics();
           emuDebugUi.drawMetaData({
-            averageRTTms, sentMessages, failedMessages, missedDraw
+            averageRTTms, sentMessages, failedMessages, missedDraw, lastFps
           });
           if (emuState.asic.wpc.inputState) {
             updateUiSwitchState(emuState.asic.wpc.inputState);
@@ -121,7 +121,6 @@ function initialiseEmu(gameEntry) {
           fpsTimes.push(timestamp);
           if (Math.abs(lastFps - fpsTimes.length) > 5) {
             lastFps = fpsTimes.length;
-            console.log('fps changed', lastFps);
             //TODO: reduce FPS
             //wpcEmuWebWorkerApi.adjustFramerate(fpsTimes.length);
           }
@@ -151,7 +150,9 @@ function saveState() {
 
 function loadState() {
   return pauseEmu()
-    .then(() => { return wpcEmuWebWorkerApi.getEmulatorRomName(); })
+    .then(() => {
+      return wpcEmuWebWorkerApi.getEmulatorRomName();
+    })
     .then((romName) => {
       const emuState = loadRam(romName);
       return wpcEmuWebWorkerApi.setEmulatorState(emuState);
@@ -212,27 +213,28 @@ function memoryMonitorPrevPage() {
 }
 
 /**
- * find data in memory
- * @param {*} value the value you are looking for
- * @param {*} encoding type of search, can be 'string', uint8, uint16
+ * Find data in emulator memory
+ * @param {Number|String} value the value you are looking for
+ * @param {String} encoding type of search, can be 'string', uint8, uint16
+ * @param {Boolean} rememberResults only uint8 supported, remembers all the positions from a previous search
  */
-function memoryFindData(value, encoding) {
-  emuDebugUi.memoryFindData(value, encoding);
+function memoryFindData(value, encoding, rememberResults = false) {
+  emuDebugUi.memoryFindData(value, encoding, rememberResults);
 }
 
 /**
- * write directly to emulator memory
- * @param {*} offset where to write
- * @param {*} value String or uint8 value to write
- * @param {*} block optional option (default is false) to persist stored data
+ * Write directly to emulator memory
+ * @param {Number} offset where to write
+ * @param {Number|String} value String or uint8 value to write
+ * @param {Boolean} block optional option (default is false) to persist stored data
  */
 function writeMemory(offset, value, block) {
   return wpcEmuWebWorkerApi.writeMemory(offset, value, block);
 }
 
 /**
- * print memory content, if its a string
- * @param {*} offset
+ * Print emulator memory content, if its a string the whole string will be shown
+ * @param {Number} offset
  */
 function memoryDumpData(offset) {
   emuDebugUi.memoryDumpData(offset);
@@ -243,7 +245,7 @@ function pauseEmu() {
   emuDebugUi.updateCanvas(null, 'paused', audioState);
 */
   soundInstance.pause();
-  cancelAnimationFrame(rafId)
+  cancelAnimationFrame(rafId);
   rafId = undefined;
   return wpcEmuWebWorkerApi.pauseEmulator();
 }