+++ client/scripts/ui/memory-monitor.js
@@ -1,12 +1,12 @@
 'use strict';
 
-import { createDrawLib } from './lib';
-import { findString, findUint8, findUint16, findUint32 } from './search';
-import { replaceNode } from '../htmlselector';
+import { createDrawLib } from './ui/lib';
+import { memoryFindData } from './ui/memorysearch';
+import { replaceNode, replaceNodeAndResize } from './htmlselector';
 
 export { getInstance };
 
-const MEM_HEIGHT = 220;
+const WINDOW_HEIGHT = 220;
 const MEM_CONTENT_X = 9;
 const MEM_CONTENT_Y = 3;
 const MEM_CONTENT_ASCII_X = 76;
@@ -15,6 +15,8 @@ const MEM_CONTENT_OFFSET_X = 2;
 const ENTRIES_HORIZONTAL = 32;
 const ENTRIES_VERTICAL = 16;
 const ENTRIES_PAGESIZE = ENTRIES_HORIZONTAL * ENTRIES_VERTICAL;
+
+const ENABLE_MEMORY_MONITOR = 'enable memory monitor first (press m)!';
 
 function getInstance(options) {
   return new MemoryMonitor(options);
@@ -28,18 +30,18 @@ class MemoryMonitor {
     this.memoryMonitorEnabled = false;
     this.page = 0;
     this.lastRamSnapshot = undefined;
+    this.memorySearchResult = undefined;
 
     const canvasMemoryElement = this._createCanvas();
     this.canvasMemory = canvasMemoryElement.getContext('2d', { alpha: true });
     replaceNode('memoryNode', canvasMemoryElement);
     this.canvasMemoryDrawLib = createDrawLib(this.canvasMemory, options.THEME);
+    this.canvasMemoryDrawLib.clear();
 
     const canvasMemoryOverlayElement = this._createCanvas();
     this.canvasMemoryOverlay = canvasMemoryOverlayElement.getContext('2d', { alpha: true });
-    replaceNode('memoryOverlayNode', canvasMemoryOverlayElement);
+    replaceNodeAndResize('memoryOverlayNode', canvasMemoryOverlayElement, WINDOW_HEIGHT);
     this.canvasMemoryOverlayDrawLib = createDrawLib(this.canvasMemoryOverlay, options.THEME);
-
-    this.canvasMemoryDrawLib.clear();
 
     // HIGHLIGHT ROWS
     for (let offset = 0; offset < ENTRIES_HORIZONTAL; offset += 2) {
@@ -83,59 +85,42 @@ class MemoryMonitor {
   _createCanvas() {
     const canvasElement = document.createElement('canvas');
     canvasElement.width = this.CANVAS_WIDTH;
-    canvasElement.height = MEM_HEIGHT;
+    canvasElement.height = WINDOW_HEIGHT;
     return canvasElement;
   }
 
   clear() {
     this.canvasMemoryOverlayDrawLib.clear();
     this.page = 0;
+    this.memorySearchResult = undefined;
     this.lastRamSnapshot = undefined;
   }
 
-  toggleMemoryView() {
-    const node = document.querySelector('#memoryMonitor');
-
-    if (this.memoryMonitorEnabled) {
-      node.style.height = '0px';
-      node.style.visibility = 'hidden';
-      this.memoryMonitorEnabled = false;
-      return;
-    }
-
-    this.memoryMonitorEnabled = true;
-    node.style.height = MEM_HEIGHT + 'px';
-    node.style.visibility = 'visible';
+  toggleView(enabled) {
+    this.memoryMonitorEnabled = enabled === true;
   }
 
-  memoryFindData(value, encoding = 'string') {
+  memoryFindData(value, encoding = 'string', rememberResults) {
     if (!this.lastRamSnapshot) {
-      console.warn('enable memory monitor first!')
-      return;
+      return console.warn(ENABLE_MEMORY_MONITOR);
     }
+    console.log('memoryFindData', { value, encoding, rememberResults });
 
-    console.log('memoryFindData', { value, encoding });
-
+    const foundOffset = memoryFindData(value, encoding, rememberResults, this.lastRamSnapshot);
     if (encoding === 'string') {
-      return findString(value, this.lastRamSnapshot);
+      foundOffset.forEach((offset) => {
+        console.log(value, encoding, 'FOUND at position', '0x' + offset.toString(16).toUpperCase());
+      });
+    } else {
+      foundOffset.forEach((offset) => {
+        console.log('0x' + value.toString(16), encoding, 'FOUND at position', '0x' + offset.toString(16).toUpperCase());
+      });
     }
-    if (encoding === 'uint8') {
-      return findUint8(value, this.lastRamSnapshot);
-    }
-    if (encoding === 'uint16') {
-      return findUint16(value, this.lastRamSnapshot);
-    }
-    if (encoding === 'uint32') {
-      return findUint32(value, this.lastRamSnapshot);
-    }
-
-    console.log('UNKNOWN_ENCODING')
   }
 
   memoryDumpData(offset) {
     if (!this.lastRamSnapshot) {
-      console.warn('enable memory monitor first!')
-      return;
+      return console.warn(ENABLE_MEMORY_MONITOR);
     }
 
     let dump = '';
@@ -191,7 +176,7 @@ class MemoryMonitor {
       );
 
       for (let offset = 0; offset < ENTRIES_HORIZONTAL; offset++) {
-        const value = this.lastRamSnapshot[ ramOffset + offset ];
+        const value = this.lastRamSnapshot[ramOffset + offset];
         // write hex value
         this.canvasMemoryOverlayDrawLib.writeHeader(
           MEM_CONTENT_X + offset * 2,